requirements.txt
google-genai>=1.26.0
google-auth>=2.40.3